# src/docs/swagger.yaml
basePath: /
definitions:
...
  carwise.ChangePasswordRequest:
    description: Change password request
    properties:
      password:
        example: NewStrongP@ss123
        type: string
      re_password:
        example: NewStrongP@ss123
        type: string
    required:
    - password
    - re_password
    type: object
...
  carwise.ResetPasswordRequest:
    description: Password reset request
    properties:
      email:
        example: john.doe@example.com
        type: string
    required:
    - email
    type: object
...
  carwise.TokenResponse:
    description: JWT token response
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  carwise.UserCreateRequest:
    description: User registration request
    properties:
      country_code:
        example: "+90"
        maxLength: 10
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 50
        minLength: 2
        type: string
      password:
        example: StrongP@ss123
        type: string
      phone_number:
        example: "5551234567"
        type: string
    required:
    - country_code
    - email
    - first_name
    - last_name
    - password
    - phone_number
    type: object
...
  carwise.UserLoginRequest:
    description: User login request
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: StrongP@ss123
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@carwise.com
    name: API Support
  description: Carwise API documentation with Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Carwise API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Password reset request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/carwise.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent successfully
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Request password reset
      tags:
      - Authentication
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user and return JWT token
      parameters:
      - description: User login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/carwise.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access token
          schema:
            $ref: '#/definitions/carwise.TokenResponse'
        "400":
          description: Validation or login error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Authentication
  /auth/logout:
    post:
      description: Logout user by blacklisting their token
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user and return JWT token
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/carwise.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns access token
          schema:
            $ref: '#/definitions/carwise.TokenResponse'
        "400":
          description: Validation or creation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using token and email
      parameters:
      - description: Reset token
        in: query
        name: token
        required: true
        type: string
      - description: User email
        in: query
        name: email
        required: true
        type: string
      - description: New password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/carwise.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Reset password
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"